refactor: separate MaterialApp
