+++ lib/app/app.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class FindItApp extends StatelessWidget {
+  const FindItApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: Scaffold(
+        body: Center(
+          child: Text('Hello World!'),
+        ),
+      ),
+    );
+  }
+}
